was added the album amd images modules

+++ src/modules/album/album.service.ts
@@ -5,20 +5,18 @@ import { AlbumServiceContract } from "./types/album.contracts";
 
 export const AlbumService: AlbumServiceContract = {
     async create(userId, data) {
-        const albums = await AlbumRepository.getAll(userId);
-
-        const customAlbums = albums.filter((a) => a.type === "custom");
-        if (customAlbums.length >= 1) {
-            throw new ConflictError("You can only have one custom album");
-        }
-
         return await AlbumRepository.create(userId, data);
+    },
+
+    async getAll(userId) {
+        return await AlbumRepository.getAll(userId);
     },
 
     async edit(userId, albumId, data) {
         const album = await AlbumRepository.findById(albumId);
 
         if (!album) throw new NotFoundError("Album");
+        if (album.userId !== userId) throw new ForbiddenError("Album does not belong to you");
         if (album.type === "system") throw new ForbiddenError("Cannot edit system album");
 
         return await AlbumRepository.edit(albumId, data);
@@ -28,19 +26,17 @@ export const AlbumService: AlbumServiceContract = {
         const album = await AlbumRepository.findById(albumId);
 
         if (!album) throw new NotFoundError("Album");
+        if (album.userId !== userId) throw new ForbiddenError("Album does not belong to you"); 
         if (album.type === "system") throw new ForbiddenError("Cannot change visibility of system album");
 
         return await AlbumRepository.editVisibility(albumId, data);
-    },
-
-    async getAll(userId) {
-        return await AlbumRepository.getAll(userId);
     },
 
     async delete(userId, albumId) {
         const album = await AlbumRepository.findById(albumId);
 
         if (!album) throw new NotFoundError("Album");
+        if (album.userId !== userId) throw new ForbiddenError("Album does not belong to you");
         if (album.type === "system") throw new ForbiddenError("Cannot delete system album");
 
         await AlbumRepository.delete(albumId);
